"Add default parameters in Snowflake constructor and update error message"
This change modified the constructor of the Snowflake class to include default values for both machineId and datacenterId. In cases where the user does not provide these parameters, the default values are generated as random numbers within their respective maximum allowable ranges. This change has made the constructor more flexible. Additionally, the error message was updated for better clarity when the inputted machineId or datacenterId exceeds maximum limit.